lib/reflect: add function Marshal
The Marshal function marshal the obj value to []byte by calling one of
the method: MarshalBinary, MarshalJSON, or MarshalText; in respective
order.

If obj implement one of the method with valid signature, it will return
(out, nil, true);
unless there is an error.

If the method signature invalid it will return (nil, err, false).

If obj is nil or none of the method exist it will return
(nil, nil, false).