lib/memfs: set the Root SysPath to the first MemFS instance on Merge
Previously, calling Merge(...), set the merged MemFS Root.SysPath to
"..".
Since we allow the TryDirect to access the file directly (if its set
to true), this may cause the file system leaks if returned MemFS set
this flag to true.

To prevent that, we set the SysPath to the first MemFS SysPath.