lib/ini: fix parsing and saving multi line variables
Previously, if INI file contains multi line variables, for example

  key = a \
  	b

The Get and saved value is "a \tb", where it should be "a b" for Get and
"a \\\n\t\b" again when saved.

This changes require refactoring how the variable's value is parsed and
stored.
A variable value is parsed and stored from character after "=" until new
line or comment as raw value, and the real value is derived by trimming
white spaces, handle escaped character and double quotes.