lib/test: use text/diff to compare strings on Assert
If both exp and got types are string and its longer than 50 chars, it
will use the text/diff.Text to show the difference between them.
The diff output is as follow,

	!!! string not matched:
	--++
	<LINE_NUM> - "<LINE_EXP>"
	<LINE_NUM> + "<LINE_GOT>"
	^<COL_NUM> - "<DELETED_STRING>"
	^<COL_NUM> + "<INSERTED_STRING>"

The "<LINE_NUM> - " print the line number in exp followed by line itself.
The "<LINE_NUM> + " print the line number in got followed by line itself.
The "^<COL_NUM> - " show the character number in exp line followed by
deleted string (or string that not exist in got).
The "^<COL_NUM> + " show the character number in got line followed by
inserted string (or string that not exist in exp).