lib/test: simplify the string diff output from Assert
In the output, instead of using %q we replace it with %s, because printing
string with double quote cause escaping and hard to read
This change may cause difference in white spaces not showed in the
terminal.

In the diff changes, only print the Old and New, without printing each
chunk.