lib/dns: simplify unpackDomainName return value for end
Previous, the unpackDomainName return the end value based on if the
domain name has compressed or not and let the caller calculated the
next index.

This changes simplify it by return the total bytes the domain name packed
inside packet, the caller only need to set index based on the returned
value.