lib/dns: add method to decode a charater-string
A character-string is expressed in one or two ways: as a contiguous set of
characters without interior spaces, or as a string beginning with a " and
ending with a ".

For contiguous string without double quotes, the following encoding are
recognized,

  - \X where X is any character other than a digit (0-9), is used to quote
    that character so that its special meaning does not apply.
    For example, "\." can be used to place a dot character in a label.
  - \DDD  where each D is a digit is the octet corresponding to the decimal
    number described by DDD.
    The resulting octet is assumed to be text and is not checked for
    special meaning.

For quoted string, inside a " delimited string any character can occur,
except for a " itself, which must be quoted using \ (back slash).

[character-string]: https://datatracker.ietf.org/doc/html/rfc1035#section-5.1