_bin: add shell script go-test-lint.sh to run test and lint sequentially
The go-test-lint.sh run Go test and if its success it will run
predefined linter, in the current directory.

Arg 1: the method or function to test, default to ".".

The linter program and its argument is derived from environment variable
GO_LINT.
If its empty, it will try the following linter in order: revive, or
golangci-lint.

To add additional arguments to go test set the environment variable
GO_TEST_ARGS.