lib/dns: fix packing, parsing, and saving MINFO resource data
Even thought the MINFO record not formally obsolete, according to
Wikipedia [1], we still need to support this for backward compatibility.

When packing the resource data length does not include total length.
When parsing, the RMailBox and EmailBox should be added the origin suffix
if its not end with dot.
When saving, the origin should be trimmed from RMailBox and EmailBox.

[1] https://en.wikipedia.org/wiki/List_of_DNS_record_types#Obsolete_record_types