lib/net: changes the PopulateQuery logic
Previously, PopulateQuery only add the passed dname if the number of dots
is greater than 0.
After inspecting the result from dig and getenv, the dots seems does not
affect the query.
For example, if we have A record for domain "kilabit", both of those tools
query name "kilabit" without adding local domain or domain in search.