lib/bytes: implement function ParseHexDump
The ParseHexDump parse the default output of [hexdump](1) utility from
parameter in back into stream of byte.

An example of default output of hexdump is

	0000000 7865 5f70 6964 2f72 0000 0000 0000 0000
	0000010 0000 0000 0000 0000 0000 0000 0000 0000
	*
	0000060 0000 0000 3030 3030 3537 0035 3030 3130

The first column is the address and the rest of the column is the data.
Each data column is 16-bit words in big-endian order, so in the above
example, the first byte would be 65, second byte is 78 and so on.
The asterisk "*" means that the address from 0000020 to 0000050 is equal to
the previous line, 0000010.

[hexdump]: https://man.archlinux.org/man/hexdump.1