lib/strings: merge lib/parser here
The first idea of parser is to provide generic parser for both bytes and
string.
After we introduce lib/parser there is not much changes to that package.
Also, since we create another Parser in lib/bytes that accept and
return token as []byte, the lib/parser is not unique anymore.

The following function/methods changes to minimize conflict in the future,

  * Lines become LinesOfFile
  * New become NewParser
  * Open become OpenForParser
  * Token become Read
  * TokenEscaped become ReadEscaped
  * TokenTrimSpace become ReadNoSpace