lib/memfs: mock the os.Stat to fix flaky tests
In the history of our tests, we got the following test that sometime
failed,

  2023/05/02 14:27:32 DirWatcher: on file changes /subdir/testfile: file does not exist
  2023/05/02 14:27:34 DirWatcher: on file changes /sub/index.html: file does not exist
  panic: test timed out after 1m0s

  ...

  goroutine 1 [chan receive]:
  github.com/shuLhan/share/lib/memfs.ExampleDirWatcher()
	share/lib/memfs/dirwatcher_example_test.go:114 +0x165f

One of possible cause is that in the DirWatcher or Watcher we have
the following condition

  // Watcher.start
  if w.node.ModTime().Equal(newInfo.ModTime()) {
     continue
  }

  // DirWatcher.processSubdirs
  if node.ModTime().Equal(newDirInfo.ModTime()) {
     continue

that result to true if the node modified in the same second as it was
created, which cause changes not propagated to qchanges.

The fix is by mocking the os.Stat by returning the FileInfo with ModTime
always increased by 1 second on every call.