lib/http: redirect path with slash if request is directory
If request path is a directory and it is not end with slash, redirect
request to location with slash to allow relative links works inside the
HTML content.

For example, a "/page/index.html" contains links href="sub.html" (where
"sub.html" is inside "/page" directory).
If request to "/page" (without end with slash) return content of
"/page/index.html", then when user click on sub.html it will request to
"/sub.html" instead of "/page/sub.html".