internal: add package asciiset
The package asciiset add new type ASCIISet, a bitmap to represent list of ASCII characters
for faster lookup.

ASCIISet is a 36-byte value, where each bit in the first 32-bytes represents the presence of
a given ASCII character in the set.
The remaining 4-bytes is a counter for the number of ASCII characters in the set.
The 128-bits of the first 16 bytes, starting with the least-significant bit of the lowest word
to the most-significant bit of the highest word, map to the full range of all 128 ASCII characters.
The 128-bits of the next 16 bytes will be zeroed, ensuring that any non-ASCII character will
be reported as not in the set.