websocket/testdata: rewrite autobahn test using container
Since the autobahn script can only run on Python 2, it become hard to
setup and run the test on distro that does not provide Python 2 anymore.
The autobahn repository recommend to use docker instead.

When testing the server, we simplify it by using make task "test-server".
The test-server task run our test server in background, and then run the
autobahn fuzzingclient from container.
Once the tests completed, we trigger the server to shutdown by sending
text frame with payload "shutdown".

When testing the client, we simplify it by using make task "test-client".
The test-client task run the autobahn fuzzingserver and then
we run our client.
Once client finished, we trigger the server to generate the reports
and cleanup the container.