lib/dns: add method AddAuthority to Message
The AddAuthority add the rr to list of Authority.
Calling this method mark the message as answer, instead of query.

If the rr is SOA, it will replace the existing record if exist and set
the flag authoritative answer (IsAA) in header to true.
If the rr is NS, it will be added only if its not exist.

It will return an error if the rr type is not SOA or NS or the size of
records in Authority is full, maximum four records.