ssh/config: convert value of CASignatureAlgorithms and Hostname to lowercase
An example of CASignatureAlgorithms is ssh-ed25519, so it make senses to
convert it to lowercase.
The value of Hostname is either a domain or IP address so it also make
senses to convert it to lowercase.
The value of User is not case insensitive since we can create user name
"Foo" and "foo" in the same system.