lib/memfs: remove embedding os.FileInfo and http.File
The original idea of embedding those interface is to guard the Node
implementation to always follow all methods in the os.FileInfo and
http.File, in case there is additional methods in the future.
But, embedding interfaces make the struct store it as field "FileInfo"
and File with "<nil>" values, which add storage cost to Node.