lib/memfs: stop all Watcher spawn by DirWatcher
Previously, the Watcher goroutine will stopped only if the file is
being deleted.
If the DirWatcher stopped manually, by calling Stop method, the
Watcher goroutine will still running in the background.

This changes record all spawned Watcher and stop it when files inside
a deleted directory or when Stop called.