lib/memfs: add method JSON to Node
The JSON method encode the Node into JSON.
This method provides an alternative to MarshalJSON with granular options,
depth and withoutModTime.

The depth set the level of childs to be encoded.
The depth=0 only encode the Root Node itself (with its childs), depth=1
encode the Root node and its subdirectories, and so on.

If withoutModTime is set to true, all of the node ModTime will not be
included in output.