lib/net: add method Read
The Read method read packet from raw connection.

If the conn parameter is nil it will return [net.ErrClosed].

The bufsize parameter set the size of buffer for each read operation,
default to 1024 if not set or invalid (less than 0 or greater than
65535).

The timeout parameter set how long to wait for data before considering
it as failed.
If its not set, less or equal to 0, it will wait forever.
If no data received and timeout is set, it will return [ErrReadTimeout].

If there is data received and connection closed at the same time, it will
return the data first without error.
The subsequent Read will return empty packet with [ErrClosed].

Signed-off-by: Shulhan <ms@kilabit.info>