lib/net: use [time.Ticker] to loop in WaitAlive
The Timeout field in net.Dialer sometimes does not have any effect.
In the for-loop, we expect that each Dial at least consume 100 ms
before returning an error.

Turns out, on several occasion, the loop run too quickly, less than the
passed timeout.
This is probably because the listening address has not been
active yet and we run it in the same machine, the OS (or Go?) return
immediately without waiting for timeout.