http/sseclient: add test for multi lines with empty line in middle
When running the EventSource client in Firefox, we found out that when
server send a multi line content with empty line in the middle, the empty
line is ignored by Firefox.

Just to make sure we test it here and the server send it as expected

  event:message\ndata:line 1\ndata:\ndata:line2\ndata:\ndata:\n\n

This case is similar with case number 3 when testing raw data.
So probably this is bug in our implementation of SSE in another project.