ssh/config: refactoring the Config merge
This changes rename method [Config.Prepend] to [Config.Merge].

The way that how the other Config merged is changed.
Instead of appending all of other's sections into the current Config,
append the other Config instance to the current instance of Config.

During [Config.Get] the top Config will be evaluated first, and then the
other Config is evaluated in order of Merge.