lib/sql: replace [http.FileSystem] with [memfs.MemFS]
Accepting the [http.FileSystem] means that the parameter can receive an
instance of [embed.FS], but in most cases, it will fail.

Case example, when we embed SQL files for migration under
"db/migration" using the "go:embed" directive,

	//go:embed db/migration/*.sql
	var DBMigrationFS embed.FS

and then call the [Migrate] function, it will not found any ".sql" files
inside the "/" directory because the files is stored under
"db/migration/" prefix (also there is no "/" when using embed.FS).