lib/play: simplify running Go code
This changes remove the field pid in the struct command that wait for
process ID, instead execute cmd with Run directly.

In the Run function, use the UnsafeRun to store temporary directory
and move the statements that writes go.mod and main.go into the method
writes of Request.
This remove unnecessary unsafeRun function.