lib/play: refactor functions as methods of type Go
One of the major issue that we previously have is the Run and Test
functions can write file in any unsafe path.
Another issue is changing the default GoVersion and Timeout is set on
the package level.

This changes introduce new type Go as the top level type that can be
instantiate with different Root, GoVersion, and Timeout.
The instance of Go then can Format, Run, or Test the Go code in their
own scope.