lib/goanalysis: package to complement "go vet"
Package goanalysis implement go static analysis using
[Analyzer] that are not included in the default "go vet", but included in
the [passes] directory, including: fieldalignment, nilness,
reflectvaluecompare, shadow, sortslice, unusedwrite, and waitgroup.

This package is not mean to be imported directly by other package
except main, like we have in [cmd/gocheck].

The rest (and previous) of the changes are affect running gocheck on it.