lib/test: implement method ExtractInput on Data
Given the path to directory `destDir`, create all of the
[test.Data.Input] with key as its file name.

If the input name contains "/", the path before the base name will be
created along with its parent as long as it is under the `destDir`.

For example, given input name "a/b/c.txt", it will create path "a/b/"
inside `destDir` first, followed by file "c.txt" inside that path.